+++ src/api/tests/create-custom-call.test.ts
@@ -1,22 +1,25 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { faker } from "@faker-js/faker"
-import { SnakeCasedPropertiesDeep } from "@thinknimble/tn-utils"
+import { SnakeCasedPropertiesDeep, objectToSnakeCase } from "@thinknimble/tn-utils"
 import { describe, expect, it, vi } from "vitest"
 import { z } from "zod"
 import { GetInferredFromRaw } from "../../utils"
 import { createApi } from "../create-api"
 import { createCustomServiceCall } from "../create-custom-call"
 import { mockedAxios } from "./mocks"
+import { CustomServiceCallPlaceholder } from "../types"
 
 describe("createCustomServiceCall", () => {
+  const inputShape = {
+    anotherInput: z.string(),
+  }
+  const outputShape = {
+    justAny: z.any(),
+  }
   const testPost = createCustomServiceCall(
     {
-      inputShape: {
-        anotherInput: z.string(),
-      },
-      outputShape: {
-        justAny: z.any(),
-      },
+      inputShape,
+      outputShape,
     },
     async ({ client, input, utils, slashEndingBaseUri }) => {
       const toApiInput = utils.toApi(input)
@@ -25,6 +28,7 @@ describe("createCustomServiceCall", () => {
       return parsed
     }
   )
+
   it("calls api with snake case", async () => {
     //arrange
     const baseUri = "callsApiWithSnakeCase"
@@ -215,6 +219,7 @@ describe("createCustomServiceCall", () => {
         return
       }
     )
+    type testNoInputNorOutput = typeof testNoInputNorOutput
     const baseUri = "noInputNorOutputOverload"
     const testApi = createApi(
       {
@@ -300,9 +305,183 @@ describe("createCustomServiceCall", () => {
       another_input: input.anotherInput,
     })
   })
+  it("passes the right filters to callback: input and output", async () => {
+    //arrange
+    const getSpy = vi.spyOn(mockedAxios, "get")
+    mockedAxios.get.mockResolvedValueOnce({
+      data: {
+        test_output: "test",
+      },
+    })
+    const callWithFilter = createCustomServiceCall(
+      {
+        inputShape: {
+          testInput: z.string(),
+        },
+        outputShape: {
+          testOutput: z.string(),
+        },
+        filtersShape: {
+          testFilter: z.string(),
+        },
+      },
+      async ({ client, slashEndingBaseUri, parsedFilters }) => {
+        const result = await client.get(slashEndingBaseUri, { params: parsedFilters })
+        return result.data
+      }
+    )
+    const baseUri = "filters"
+    const filters = {
+      testFilter: "myFilter",
+    }
+    const api = createApi(
+      {
+        client: mockedAxios,
+        baseUri,
+      },
+      {
+        callWithFilter,
+      }
+    )
+    //act
+    await api.csc.callWithFilter({
+      input: { testInput: "testInput" },
+      filters,
+    })
+    //assert
+    expect(getSpy).toHaveBeenCalledWith(`${baseUri}/`, {
+      params: objectToSnakeCase(filters),
+    })
+  })
+  it("passes the right filters to callback: only output", async () => {
+    //arrange
+    const getSpy = vi.spyOn(mockedAxios, "get")
+    mockedAxios.get.mockResolvedValueOnce({
+      data: {
+        test_output: "test",
+      },
+    })
+    const callWithFilter = createCustomServiceCall(
+      {
+        outputShape: {
+          testOutput: z.string(),
+        },
+        filtersShape: {
+          testFilter: z.string(),
+        },
+      },
+      async ({ client, slashEndingBaseUri, parsedFilters }) => {
+        const result = await client.get(slashEndingBaseUri, { params: parsedFilters })
+        return result.data
+      }
+    )
+    const baseUri = "filters"
+    const filters = {
+      testFilter: "myFilter",
+    }
+    const api = createApi(
+      {
+        client: mockedAxios,
+        baseUri,
+      },
+      {
+        callWithFilter,
+      }
+    )
+    //act
+    await api.csc.callWithFilter({
+      filters,
+    })
+    //assert
+    expect(getSpy).toHaveBeenCalledWith(`${baseUri}/`, {
+      params: objectToSnakeCase(filters),
+    })
+  })
+  it("does not error on not passing filters", async () => {
+    //arrange
+    const getSpy = vi.spyOn(mockedAxios, "get")
+    mockedAxios.get.mockResolvedValueOnce({
+      data: {
+        test_output: "test",
+      },
+    })
+    const callWithFilter = createCustomServiceCall(
+      {
+        outputShape: {
+          testOutput: z.string(),
+        },
+        filtersShape: {
+          testFilter: z.string(),
+        },
+      },
+      async ({ client, slashEndingBaseUri, parsedFilters }) => {
+        const result = await client.get(slashEndingBaseUri, { params: parsedFilters })
+        return result.data
+      }
+    )
+    const baseUri = "filters"
+    const api = createApi(
+      {
+        client: mockedAxios,
+        baseUri,
+      },
+      {
+        callWithFilter,
+      }
+    )
+    //act
+    await api.csc.callWithFilter()
+    //assert
+    expect(getSpy).toHaveBeenCalledWith(`${baseUri}/`, {
+      params: undefined,
+    })
+  })
+  it("should TS error if user passes filtersShape but no output", async () => {
+    //arrange
+    const getSpy = vi.spyOn(mockedAxios, "get")
+    mockedAxios.get.mockResolvedValueOnce({
+      data: {
+        test_output: "test",
+      },
+    })
+    const callWithFilter = createCustomServiceCall(
+      {
+        //@ts-expect-error cannot pass filter shape if there is no output shape
+        filtersShape: {
+          testFilter: z.string(),
+        },
+      },
+      async () => {
+        //no-op
+      }
+    )
+  })
+  it("Should not allow filters if there is no output (just input)", async () => {
+    //arrange
+    const getSpy = vi.spyOn(mockedAxios, "get")
+    mockedAxios.get.mockResolvedValueOnce({
+      data: {
+        test_output: "test",
+      },
+    })
+    //@ts-expect-error should not allow to create with filter and no outputshape
+    const callWithFilter = createCustomServiceCall(
+      {
+        inputShape: {
+          testInput: z.string(),
+        },
+        filtersShape: {
+          testFilter: z.string(),
+        },
+      },
+      async ({ client, slashEndingBaseUri }) => {
+        //no-op
+      }
+    )
+  })
 
   describe("standAlone call", () => {
-    it("can convert to stand alone call and calls with right parameters: input object output object", async () => {
+    it("calls with right parameters: input object output object", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = {
@@ -323,7 +502,13 @@ describe("createCustomServiceCall", () => {
           },
         },
         name: callName,
-        cb: async ({ client, utils, input }) => {
+        cb: async ({
+          client,
+          utils,
+          input,
+          //@ts-expect-error should not yield uri since none was declared
+          slashEndingBaseUri,
+        }) => {
           const res = await client.post(`${callName}/`, utils.toApi(input))
           return utils.fromApi(res.data)
         },
@@ -336,7 +521,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toEqual({ testData: mockResult.test_data })
     })
 
-    it("can convert to stand alone call and calls with right parameters: input object output primitive", async () => {
+    it("calls with right parameters: input object output primitive", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = [faker.datatype.string(), faker.datatype.string()]
@@ -366,7 +551,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toEqual(mockResult)
     })
 
-    it("can convert to stand alone call and calls with right parameters: input object output void", async () => {
+    it("calls with right parameters: input object output void", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = {
@@ -396,7 +581,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toBeUndefined()
     })
 
-    it("can convert to stand alone call and calls with right parameters: input void output object", async () => {
+    it("calls with right parameters: input void output object", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = {
@@ -426,7 +611,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toEqual({ testData: mockResult.test_data })
     })
 
-    it("can convert to stand alone call and calls with right parameters: input void output void", async () => {
+    it("calls with right parameters: input void output void", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = {
@@ -450,7 +635,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toBeUndefined()
     })
 
-    it("can convert to stand alone call and calls with right parameters: input void output primitive", async () => {
+    it("calls with right parameters: input void output primitive", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = Array.from({ length: 5 })
@@ -481,7 +666,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toEqual(mockResult)
     })
 
-    it("can convert to stand alone call and calls with right parameters: input primitive output object", async () => {
+    it("calls with right parameters: input primitive output object", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = {
@@ -513,7 +698,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toEqual({ testData: mockResult.test_data })
     })
 
-    it("can convert to stand alone call and calls with right parameters: input primitive output void", async () => {
+    it("calls with right parameters: input primitive output void", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = {
@@ -541,7 +726,7 @@ describe("createCustomServiceCall", () => {
       expect(result).toBeUndefined()
     })
 
-    it("can convert to stand alone call and calls with right parameters: input primitive output primitive", async () => {
+    it("calls with right parameters: input primitive output primitive", async () => {
       //arrange
       const postSpy = vi.spyOn(mockedAxios, "post")
       const mockResult = Array.from({ length: 5 })
@@ -566,12 +751,251 @@ describe("createCustomServiceCall", () => {
           return utils.fromApi(res.data)
         },
       })
-
       const testStrInput = faker.name.firstName()
       //act
       const result = await testStandAloneCall(testStrInput)
       //assert
       expect(postSpy).toHaveBeenLastCalledWith(`${callName}/${testStrInput}/`)
+      expect(result).toEqual(mockResult)
+    })
+
+    // //-----
+    it("calls with right parameters and filters: input object output object", async () => {
+      //arrange
+      const postSpy = vi.spyOn(mockedAxios, "post")
+      const mockResult = {
+        test_data: "testData",
+      }
+      mockedAxios.post.mockResolvedValueOnce({
+        data: mockResult,
+      })
+      const callName = "testStandAloneCall"
+      const testStandAloneCall = createCustomServiceCall.standAlone({
+        client: mockedAxios,
+        models: {
+          outputShape: {
+            testData: z.string(),
+          },
+          inputShape: {
+            testInput: z.number(),
+          },
+          filtersShape: {
+            testFilter: z.string(),
+          },
+        },
+        name: callName,
+        cb: async ({ client, utils, input, parsedFilters }) => {
+          const res = await client.post(`${callName}/`, utils.toApi(input), {
+            params: parsedFilters,
+          })
+          return utils.fromApi(res.data)
+        },
+      })
+      const testNumberInput = faker.datatype.number()
+      const testFilter = faker.datatype.string(5)
+      //act
+      const result = await testStandAloneCall({
+        input: { testInput: testNumberInput },
+        filters: { testFilter: testFilter },
+      })
+      //assert
+      expect(postSpy).toHaveBeenLastCalledWith(
+        `${callName}/`,
+        { test_input: testNumberInput },
+        {
+          params: {
+            test_filter: testFilter,
+          },
+        }
+      )
+      expect(result).toEqual({ testData: mockResult.test_data })
+    })
+
+    it("calls with right parameters and filters: input object output primitive", async () => {
+      //arrange
+      const postSpy = vi.spyOn(mockedAxios, "post")
+      const mockResult = [faker.datatype.string(), faker.datatype.string()]
+      mockedAxios.post.mockResolvedValueOnce({
+        data: mockResult,
+      })
+      const callName = "testStandAloneCall"
+      const testStandAloneCall = createCustomServiceCall.standAlone({
+        client: mockedAxios,
+        models: {
+          outputShape: z.string().array(),
+          inputShape: {
+            testInput: z.number(),
+          },
+          filtersShape: {
+            testFilter: z.string(),
+          },
+        },
+        name: callName,
+        cb: async ({ client, utils, input, parsedFilters }) => {
+          const res = await client.post(`${callName}/`, utils.toApi(input), { params: parsedFilters })
+          return utils.fromApi(res.data)
+        },
+      })
+      const testNumberInput = faker.datatype.number()
+      const testFilter = faker.datatype.string()
+      //act
+      const result = await testStandAloneCall({ input: { testInput: testNumberInput }, filters: { testFilter } })
+      //assert
+      expect(postSpy).toHaveBeenLastCalledWith(
+        `${callName}/`,
+        { test_input: testNumberInput },
+        { params: { test_filter: testFilter } }
+      )
+      expect(result).toEqual(mockResult)
+    })
+
+    it("calls with right parameters and filters: input void output object", async () => {
+      //arrange
+      const postSpy = vi.spyOn(mockedAxios, "post")
+      const mockResult = {
+        test_data: "testData",
+      }
+      mockedAxios.post.mockResolvedValue({
+        data: mockResult,
+      })
+      const callName = "testStandAloneCall"
+      const testStandAloneCall = createCustomServiceCall.standAlone({
+        client: mockedAxios,
+        models: {
+          outputShape: {
+            testData: z.string(),
+          },
+          filtersShape: {
+            testFilter: z.string(),
+          },
+        },
+        name: callName,
+        cb: async ({ client, utils, parsedFilters }) => {
+          const res = await client.post(`${callName}/`, undefined, { params: parsedFilters })
+          return utils.fromApi(res.data)
+        },
+      })
+      const testFilter = faker.datatype.string(5)
+      //act
+      const result = await testStandAloneCall({ filters: { testFilter: testFilter } })
+      //assert
+      expect(postSpy).toHaveBeenLastCalledWith(`${callName}/`, undefined, {
+        params: {
+          test_filter: testFilter,
+        },
+      })
+      expect(result).toEqual({ testData: mockResult.test_data })
+      mockedAxios.post.mockClear()
+    })
+
+    it("calls with right parameters and filters: input void output primitive", async () => {
+      //arrange
+      const postSpy = vi.spyOn(mockedAxios, "post")
+      const mockResult = Array.from({ length: 5 })
+        .fill(undefined)
+        .map(() => {
+          return faker.datatype.number()
+        })
+      mockedAxios.post.mockResolvedValueOnce({
+        data: mockResult,
+      })
+      const callName = "testStandAloneCall"
+      const testStandAloneCall = createCustomServiceCall.standAlone({
+        client: mockedAxios,
+        models: {
+          outputShape: z.string().array(),
+          filtersShape: {
+            testFilter: z.string(),
+          },
+        },
+        name: callName,
+        cb: async ({ client, parsedFilters }) => {
+          const res = await client.post(`${callName}/`, undefined, { params: parsedFilters })
+          return res.data
+        },
+      })
+      const testFilter = faker.datatype.string()
+      //act
+      const result = await testStandAloneCall({ filters: { testFilter: testFilter } })
+      //assert
+      expect(postSpy).toHaveBeenLastCalledWith(`${callName}/`, undefined, { params: { test_filter: testFilter } })
+      expect(result).toEqual(mockResult)
+    })
+
+    it("calls with right parameters and filters: input primitive output object", async () => {
+      //arrange
+      const postSpy = vi.spyOn(mockedAxios, "post")
+      const mockResult = {
+        test_data: "testData",
+      }
+      mockedAxios.post.mockResolvedValueOnce({
+        data: mockResult,
+      })
+      const callName = "testStandAloneCall"
+      const testStandAloneCall = createCustomServiceCall.standAlone({
+        client: mockedAxios,
+        models: {
+          outputShape: {
+            testData: z.string(),
+          },
+          inputShape: z.string(),
+          filtersShape: {
+            testFilter: z.string(),
+          },
+        },
+        name: callName,
+        cb: async ({ client, utils, input, parsedFilters }) => {
+          const res = await client.post(`${callName}/${input}/`, undefined, { params: parsedFilters })
+          return utils.fromApi(res.data)
+        },
+      })
+      const testStrInput = faker.name.firstName()
+      const testFilter = faker.datatype.string(5)
+      //act
+      const result = await testStandAloneCall({ input: testStrInput, filters: { testFilter } })
+      //assert
+      expect(postSpy).toHaveBeenLastCalledWith(`${callName}/${testStrInput}/`, undefined, {
+        params: {
+          test_filter: testFilter,
+        },
+      })
+      expect(result).toEqual({ testData: mockResult.test_data })
+    })
+
+    it("calls with right parameters and filters: input primitive output primitive", async () => {
+      //arrange
+      const postSpy = vi.spyOn(mockedAxios, "post")
+      const mockResult = Array.from({ length: 5 })
+        .fill(undefined)
+        .map(() => {
+          return faker.datatype.string()
+        })
+      mockedAxios.post.mockResolvedValueOnce({
+        data: mockResult,
+      })
+      const callName = "testStandAloneCall"
+      const testStandAloneCall = createCustomServiceCall.standAlone({
+        client: mockedAxios,
+        models: {
+          inputShape: z.string(),
+          outputShape: z.string().array(),
+          filtersShape: { testFilter: z.string() },
+        },
+        name: callName,
+        cb: async ({ client, input, utils, parsedFilters }) => {
+          const res = await client.post(`${callName}/${input}/`, undefined, { params: parsedFilters })
+          //
+          return utils.fromApi(res.data)
+        },
+      })
+      const testStrInput = faker.name.firstName()
+      const testFilter = faker.datatype.string()
+      //act
+      const result = await testStandAloneCall({ input: testStrInput, filters: { testFilter } })
+      //assert
+      expect(postSpy).toHaveBeenLastCalledWith(`${callName}/${testStrInput}/`, undefined, {
+        params: { test_filter: testFilter },
+      })
       expect(result).toEqual(mockResult)
     })
   })